conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.3
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
attrs=24.3.0=py39hca03da5_0
blas=1.0=openblas
brotli-python=1.0.9=py39h313beb8_9
ca-certificates=2025.4.26=hbd8a1cb_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.2.1=py39h48ca7d4_1
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.8.11=py39h313beb8_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.2.0=pyhd8ed1ab_0
filelock=3.17.0=py39hca03da5_0
fonttools=4.55.3=py39h80987f9_0
freetype=2.13.3=h47d26ad_0
fsspec=2024.12.0=py39hca03da5_0
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py39h5c1b81f_0
importlib-metadata=8.6.1=pyha770c72_0
importlib_resources=6.4.0=py39hca03da5_0
iopath=0.1.10=pypi_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.18.1=pyh707e725_3
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=py39hca03da5_0
jpeg=9e=h80987f9_3
jsonschema=4.23.0=py39hca03da5_0
jsonschema-specifications=2023.7.1=py39hca03da5_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.4=py39h313beb8_0
lcms2=2.16=he26ebf3_1
lerc=4.0.0=h313beb8_0
libabseil=20240116.2=cxx17_h313beb8_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.22=h80987f9_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libprotobuf=4.25.3=h514c7bf_0
libsodium=1.0.18=h27ca646_1
libtiff=4.7.0=h91aec0a_0
libwebp-base=1.3.2=h80987f9_1
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
markupsafe=3.0.2=py39h80987f9_0
matplotlib=3.9.2=py39hca03da5_1
matplotlib-base=3.9.2=py39h7ef442a_1
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py39hca03da5_0
narwhals=1.31.0=py39hca03da5_1
nbformat=5.10.4=py39hca03da5_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.2.1=py39hca03da5_0
numpy=1.26.4=py39h3b2db8e_0
numpy-base=1.26.4=py39ha9811e2_0
openjpeg=2.5.2=hba36e21_1
openssl=3.5.0=h81ee809_0
packaging=24.2=py39hca03da5_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=py39h41ba818_1
pip=25.1=pyhc872135_1
platformdirs=4.3.7=pyh29332c3_0
plotly=6.0.1=py39hd096484_0
portalocker=3.1.1=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
psutil=5.8.0=py39h5161555_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pybind11=2.13.6=py39h48ca7d4_1
pybind11-abi=4=hd3eb1b0_1
pybind11-global=2.13.6=py39h48ca7d4_1
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.0=py39hca03da5_0
python=3.9.21=hb885b13_1
python-dateutil=2.9.0post0=py39hca03da5_2
python-fastjsonschema=2.20.0=py39hca03da5_0
python_abi=3.9=2_cp39
pytorch=2.3.1=cpu_py39h353dffe_0
pytorch3d=0.7.8=pypi_0
pyzmq=26.2.0=py39h313beb8_0
readline=8.2=h1a28f6b_0
referencing=0.30.2=py39hca03da5_0
rpds-py=0.22.3=py39h2aea54e_0
scipy=1.13.1=py39hd336fd7_1
setuptools=78.1.1=py39hca03da5_0
six=1.17.0=py39hca03da5_0
sleef=3.5.1=h80987f9_2
sqlite=3.45.3=h80987f9_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.3=py39hca03da5_1
tk=8.6.14=h6ba3021_0
torchvision=0.18.1=cpu_py39hb6554cd_0
tornado=6.4.2=py39h80987f9_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=py39hca03da5_0
tzdata=2025b=h04d1e81_0
unicodedata2=15.1.0=py39h80987f9_1
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py39hca03da5_0
xz=5.6.4=h80987f9_1
zeromq=4.3.5=h313beb8_0
zipp=3.21.0=py39hca03da5_0
zlib=1.2.13=h18a0788_1
zstd=1.5.6=hfb09047_0
